+++ cmake/ThirdPartyMetis.cmake
@@ -31,6 +31,16 @@ IF (NEKTAR_USE_METIS)
     IF (THIRDPARTY_BUILD_METIS)
         THIRDPARTY_LIBRARY(METIS_LIBRARY STATIC metis DESCRIPTION "Metis library")
 
+        IF(CMAKE_VERSION VERSION_GREATER_EQUAL "4.0")
+            UNSET(PATCH CACHE)
+            FIND_PROGRAM(PATCH patch)
+            IF(NOT PATCH)
+	        MESSAGE(FATAL_ERROR
+	            "'patch' tool for modifying files not found. Cannot build metis.")
+            ENDIF()
+	    SET(METIS_PATCH_COMMAND ${PATCH} -p1 -f < ${PROJECT_SOURCE_DIR}/cmake/thirdparty-patches/metis-5.1.0-cmake4.0.patch)
+        ENDIF()
+
         EXTERNALPROJECT_ADD(
             metis-5.1.0
             PREFIX ${TPSRC}
@@ -43,6 +53,7 @@ IF (NEKTAR_USE_METIS)
             TMP_DIR ${TPBUILD}/metis-5.1.0-tmp
             INSTALL_DIR ${TPDIST}
             BUILD_BYPRODUCTS ${METIS_LIBRARY}
+	    PATCH_COMMAND ${METIS_PATCH_COMMAND}
             CONFIGURE_COMMAND ${CMAKE_COMMAND}
                 ${NEKTAR_EXTERNAL_PROJECT_CMAKE_GENERATOR_ARGS}
                 -DCMAKE_C_COMPILER:FILEPATH=${CMAKE_C_COMPILER}