+++ CHANGELOG.md
@@ -45,6 +45,7 @@ v5.10.0
 - Fix Boost, FFTW, VTK, and Zlib thirdparty compilation on macos, fix serial PETsc thirdparty compilation, fix OCE depedencies (!2538)
 - Add scaling factor to limit the Newton step (!2533)
 - Acctivate the inexact forcing in the Newton process (!2534)
+- Fix Arpack, Cwipi, GSMPI, Lapack, LST, Metis, OCE, TetGen, Triangle, TinyXML, and Zlib compilation for CMake 4.0 and above (!2559)
 
 **IncNavierStokesSolver**
 - Added an option to process Reynolds Stress fields at a higher polynomial order (!2303)